Refactor SpanSampleTable into its own component
